restructure repo
This does not fully set the repo up but has the new structure
Related https://github.com/capeprivacy/planning/issues/1277
