+++ cape_privacy/pandas/coordinator/client_test.py
@@ -3,7 +3,7 @@ import json
 import pytest
 import responses
 
-from cape.auth.api_token import create_api_token
+from cape_privacy.pandas.auth.api_token import create_api_token
 
 from .client import Client
 from .client import GraphQLException
