+++ cape_privacy/spark/setup/cluster.sh
