+++ requirements.txt
@@ -17,3 +17,4 @@ flake8-isort==3.0.0
 isort==4.3.21
 coverage==5.1
 validators==0.15.0
+pyspark==3.0.0
