--- spark/setup.py
@@ -1,6 +1,0 @@
-import setuptools
-
-setuptools.setup(
-    name="cape_spark",
-    version="0.0.0",
-    packages=setuptools.find_packages())
